A RandomQuote-Generator
This Random Quote Generator Freecodecamp project made by wemdavis allows users to generate random quotes and share them on Twitter. Each time you click the "New Quote" button, a new random quote will be displayed.

To generate a new quote, you can click the "New Quote" button on the page. This will trigger the application to fetch a new random quote from the available quotes and display it on the screen.

If you find a quote that you like and want to share it on Twitter, you can click the "Tweet Quote" button. This will open a new window or tab with a pre-filled tweet containing the quote and the author's name. You can then review and edit the tweet before posting it on your Twitter account.